coupon-service/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-multipart
requests
prometheus-fastapi-instrumentator
python-dotenv
